llm_retrieval_system/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# LLM and AI
google-generativeai>=0.3.0
langchain==0.0.335
tiktoken>=0.8.0

# Vector Store
pinecone>=3.0.0
sentence-transformers==2.2.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-email==0.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3
httpx==0.25.2
aiofiles==23.2.0
backoff==2.2.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring and Logging
python-json-logger==2.0.7